requirements.txt
bs4
flask
gunicorn
python-dotenv
requests
